requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Groq SDK
groq==0.4.1

# Web framework
fastapi==0.109.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6

# Vector database
faiss-cpu==1.9.0
# For GPU: faiss-gpu==1.7.4

# NLP & Embeddings
sentence-transformers==2.3.1
tiktoken==0.5.2

# Code parsing
tree-sitter==0.20.4
#tree-sitter-python==0.20.4
#tree-sitter-javascript==0.20.3
#tree-sitter-java==0.20.2
#tree-sitter-cpp==0.20.3

# Git operations
GitPython==3.1.40

# GitHub API
PyGithub==2.1.1

# Jira API
jira==3.6.0

# Confluence API
atlassian-python-api==3.41.4

# HTTP requests
httpx==0.26.0
requests==2.31.0

# Task queue (optional)
celery[redis]==5.3.4
redis==4.6.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Utilities
tenacity==8.2.3  # Retry logic
python-json-logger==2.0.7
prometheus-client==0.19.0

# CLI
click==8.1.7
rich==13.7.0
typer==0.9.0

# Testing
pytest==7.4.3
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.26.0  # For async testing

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0

# Streamlit UI (optional)
streamlit==1.29.0

# Audio processing (for Whisper)
soundfile==0.12.1
librosa==0.10.1

# Image processing (for diagram extraction)
Pillow==10.2.0
pdf2image==1.17.0

# Document parsing
python-docx==1.1.0
PyPDF2==3.0.1
markdown==3.5.1

# Monitoring
sentry-sdk==1.39.2

# API documentation
mkdocs==1.5.3
mkdocs-material==9.5.3
